+++ packages/backend/src/server/api/endpoints/admin/update-proxy-account.ts
@@ -4,6 +4,7 @@
  */
 
 import { Injectable } from '@nestjs/common';
+import * as v from 'valibot';
 import { Endpoint } from '@/server/api/endpoint-base.js';
 import {
 	descriptionSchema,
@@ -26,12 +27,9 @@ export const meta = {
 	},
 } as const;
 
-export const paramDef = {
-	type: 'object',
-	properties: {
-		description: { ...descriptionSchema, nullable: true },
-	},
-} as const;
+export const paramDef = v.object({
+	description: v.nullish(descriptionSchema),
+});
 
 @Injectable()
 export default class extends Endpoint<typeof meta, typeof paramDef> { // eslint-disable-line import/no-default-export